+++ Examples/PingTool/Assets/Scripts/PlayerMotion.cs
@@ -2,7 +2,7 @@ using Unity.Netcode.Components;
 using UnityEngine;
 
 #if UNITY_EDITOR
-using Unity.Netcode.Editor;
+using Unity.Netcode.GameObjects.Editor;
 using UnityEditor;
 /// <summary>
 /// The custom editor for the <see cref="MoverScript"/> component.