+++ com.unity.netcode.gameobjects/Editor/Analytics/NetworkManagerAnalyticsHandler.cs
@@ -1,7 +1,7 @@
 #if UNITY_EDITOR
 using UnityEngine.Analytics;
 
-namespace Unity.Netcode.Editor
+namespace Unity.Netcode.GameObjects.Editor
 {
     [AnalyticInfo("NGO_NetworkManager", "unity.netcode", 5, 100, 1000)]
     internal class NetworkManagerAnalyticsHandler : AnalyticsHandler<NetworkManagerAnalytics>