+++ com.unity.netcode.gameobjects/Editor/CodeGen/PostProcessorReflectionImporter.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using System.Reflection;
 using Mono.Cecil;
 
-namespace Unity.Netcode.Editor.CodeGen
+namespace Unity.Netcode.GameObjects.Editor.CodeGen
 {
     internal class PostProcessorReflectionImporter : DefaultReflectionImporter
     {