+++ com.unity.netcode.gameobjects/Editor/Configuration/NetcodeSettingsProvider.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 using Directory = UnityEngine.Windows.Directory;
 using File = UnityEngine.Windows.File;
 
-namespace Unity.Netcode.Editor.Configuration
+namespace Unity.Netcode.GameObjects.Editor.Configuration
 {
     internal static class NetcodeSettingsProvider
     {