+++ com.unity.netcode.gameobjects/Editor/NetworkManagerHelper.cs
@@ -1,12 +1,12 @@
 using System.Collections.Generic;
 using System.Linq;
-using Unity.Netcode.Editor.Configuration;
+using Unity.Netcode.GameObjects.Editor.Configuration;
 using Unity.Netcode.Logging;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
-namespace Unity.Netcode.Editor
+namespace Unity.Netcode.GameObjects.Editor
 {
 #if UNITY_EDITOR
     /// <summary>