+++ com.unity.netcode.gameobjects/Editor/NetworkObjectEditor.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using UnityEditor;
 using UnityEngine;
 
-namespace Unity.Netcode.Editor
+namespace Unity.Netcode.GameObjects.Editor
 {
     /// <summary>
     /// The <see cref="CustomEditor"/> for <see cref="NetworkObject"/>