+++ com.unity.netcode.gameobjects/Tests/Editor/Messaging/MessageSendingTests.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 using UnityEngine.TestTools;
 using Random = System.Random;
 
-namespace Unity.Netcode.EditorTests
+namespace Unity.Netcode.GameObjects.EditorTests
 {
     internal class MessageSendingTests
     {